+++ src/main/java/de/hpi/julianweise/slave/partition/column/sorted/ADBLongColumnSorted.java
@@ -8,7 +8,7 @@ import lombok.NoArgsConstructor;
 public class ADBLongColumnSorted extends ADBColumnSorted {
     private long[] values;
 
-    public ADBLongColumnSorted(int nodeId, int partitionId, long[] values, short[] originalIndices) {
+    public ADBLongColumnSorted(int nodeId, int partitionId, long[] values, int[] originalIndices) {
         super(nodeId, partitionId, originalIndices);
         this.values = values;
     }
