+++ src/test/java/de/hpi/julianweise/slave/query/join/ADBPartitionJoinExecutorTest.java
@@ -94,10 +94,10 @@ public class ADBPartitionJoinExecutorTest {
         TestProbe<ADBPartitionManager.Command> rightPartitionManager = testKit.createTestProbe();
 
         Map<String, ADBColumnSorted> leftSideAttributes = new Object2ObjectOpenHashMap<>();
-        leftSideAttributes.put("aInteger", new ADBIntColumnSorted(0, lPartitionId, new int[]{1, 3, 5}, new short[]{0, 1, 2}));
+        leftSideAttributes.put("aInteger", new ADBIntColumnSorted(0, lPartitionId, new int[]{1, 3, 5}, new int[]{0, 1, 2}));
 
         Map<String, ADBColumnSorted> rightSideAttributes = new Object2ObjectOpenHashMap<>();
-        rightSideAttributes.put("aInteger", new ADBIntColumnSorted(0, lPartitionId, new int[]{3, 5, 9}, new short[]{0, 1, 2}));
+        rightSideAttributes.put("aInteger", new ADBIntColumnSorted(0, lPartitionId, new int[]{3, 5, 9}, new int[]{0, 1, 2}));
 
         ADBJoinQuery joinQuery = new ADBJoinQuery();
         joinQuery.addPredicate(new ADBJoinQueryPredicate(ADBQueryTerm.RelationalOperator.EQUALITY, "aInteger", "aInteger"));
@@ -174,10 +174,10 @@ public class ADBPartitionJoinExecutorTest {
         TestProbe<ADBPartitionManager.Command> rightPartitionManager = testKit.createTestProbe();
 
         Map<String, ADBColumnSorted> rightSideAttributes = new Object2ObjectOpenHashMap<>();
-        rightSideAttributes.put("eDouble", new ADBDoubleColumnSorted(0, lPartitionId, new double[]{1, 3, 5}, new short[]{0, 1, 2}));
+        rightSideAttributes.put("eDouble", new ADBDoubleColumnSorted(0, lPartitionId, new double[]{1, 3, 5}, new int[]{0, 1, 2}));
 
         Map<String, ADBColumnSorted> leftSideAttributes = new Object2ObjectOpenHashMap<>();
-        leftSideAttributes.put("eDouble", new ADBDoubleColumnSorted(0, lPartitionId, new double[]{3, 5, 9}, new short[]{0, 1, 2}));
+        leftSideAttributes.put("eDouble", new ADBDoubleColumnSorted(0, lPartitionId, new double[]{3, 5, 9}, new int[]{0, 1, 2}));
 
         ADBJoinQuery joinQuery = new ADBJoinQuery();
         joinQuery.addPredicate(new ADBJoinQueryPredicate(ADBQueryTerm.RelationalOperator.LESS, "eDouble", "eDouble"));
@@ -249,10 +249,10 @@ public class ADBPartitionJoinExecutorTest {
         TestProbe<ADBPartitionManager.Command> rightPartitionManager = testKit.createTestProbe();
 
         Map<String, ADBColumnSorted> rightSideAttributes = new Object2ObjectOpenHashMap<>();
-        rightSideAttributes.put("aInteger", new ADBIntColumnSorted(0, lPartitionId, new int[]{1, 3, 5}, new short[]{0, 1, 2}));
+        rightSideAttributes.put("aInteger", new ADBIntColumnSorted(0, lPartitionId, new int[]{1, 3, 5}, new int[]{0, 1, 2}));
 
         Map<String, ADBColumnSorted> leftSideAttributes = new Object2ObjectOpenHashMap<>();
-        leftSideAttributes.put("aInteger", new ADBIntColumnSorted(0, lPartitionId, new int[]{3, 5, 9}, new short[]{0, 1, 2}));
+        leftSideAttributes.put("aInteger", new ADBIntColumnSorted(0, lPartitionId, new int[]{3, 5, 9}, new int[]{0, 1, 2}));
 
         ADBJoinQuery joinQuery = new ADBJoinQuery();
         joinQuery.addPredicate(new ADBJoinQueryPredicate(ADBQueryTerm.RelationalOperator.LESS, "aInteger", "aInteger"));
@@ -324,10 +324,10 @@ public class ADBPartitionJoinExecutorTest {
         TestProbe<ADBPartitionManager.Command> rightPartitionManager = testKit.createTestProbe();
 
         Map<String, ADBColumnSorted> leftSideAttributes = new Object2ObjectOpenHashMap<>();
-        leftSideAttributes.put("aInteger", new ADBIntColumnSorted(0, lPartitionId, new int[]{1, 3, 5}, new short[]{0, 1, 2}));
+        leftSideAttributes.put("aInteger", new ADBIntColumnSorted(0, lPartitionId, new int[]{1, 3, 5}, new int[]{0, 1, 2}));
 
         Map<String, ADBColumnSorted> rightSideAttributes = new Object2ObjectOpenHashMap<>();
-        rightSideAttributes.put("bInteger", new ADBIntColumnSorted(0, lPartitionId, new int[]{3, 5, 9}, new short[]{0, 1, 2}));
+        rightSideAttributes.put("bInteger", new ADBIntColumnSorted(0, lPartitionId, new int[]{3, 5, 9}, new int[]{0, 1, 2}));
 
         ADBJoinQuery joinQuery = new ADBJoinQuery();
         joinQuery.addPredicate(new ADBJoinQueryPredicate(ADBQueryTerm.RelationalOperator.GREATER, "aInteger", "bInteger"));
@@ -396,10 +396,10 @@ public class ADBPartitionJoinExecutorTest {
         TestProbe<ADBPartitionManager.Command> rightPartitionManager = testKit.createTestProbe();
 
         Map<String, ADBColumnSorted> leftSideAttributes = new Object2ObjectOpenHashMap<>();
-        leftSideAttributes.put("aInteger", new ADBIntColumnSorted(0, lPartitionId, new int[]{1, 3, 6}, new short[]{0, 1, 2}));
+        leftSideAttributes.put("aInteger", new ADBIntColumnSorted(0, lPartitionId, new int[]{1, 3, 6}, new int[]{0, 1, 2}));
 
         Map<String, ADBColumnSorted> rightSideAttributes = new Object2ObjectOpenHashMap<>();
-        rightSideAttributes.put("aInteger", new ADBIntColumnSorted(0, lPartitionId, new int[]{3, 5, 9}, new short[]{0, 1, 2}));
+        rightSideAttributes.put("aInteger", new ADBIntColumnSorted(0, lPartitionId, new int[]{3, 5, 9}, new int[]{0, 1, 2}));
 
 
         ADBJoinQuery joinQuery = new ADBJoinQuery();
